requirements.txt

boto3>=1.26.0
opencv-python>=4.7.0
numpy>=1.22.0
python-dotenv>=0.21.0